declare ShellError and ShellPromise in the right place
Previously they were declared at `Bun.ShellError` and `Bun.ShellPromise`

Also moves shell to shell.d.ts

Also adds a few type-level test cases for shell APIs
